requirements.txt
colorlog>=6.9.0
dataclasses-json>=0.6.7
hydra-core>=1.3.2
mujoco>=3.3.5
numpy>=2.3.2
omegaconf>=2.3.0
pillow>=11.3.0
psutil>=7.0.0
pydantic>=2.11.7
pygame>=2.6.1
pynput>=1.8.1
scipy>=1.16.1
typing-extensions>=4.14.1